Move all hunk diff code away into a separate module
